Revert "Refactor directory `third_party` to `assets`"
This reverts commit 4cf58ee2b48dbacf56774d58ddca0ff7449da39f.
