+++ packages/extension/src/view/devtools/tabs.tsx
@@ -48,10 +48,10 @@ import {
  */
 // @ts-ignore
 // eslint-disable-next-line import/no-relative-packages
-import SettingsTab from '../../../../../assets/icons/settings-tab.svg';
+import SettingsTab from '../../../../../third_party/icons/settings-tab.svg';
 // @ts-ignore
 // eslint-disable-next-line import/no-relative-packages
-import SettingsTabWhite from '../../../../../assets/icons/settings-tab-white.svg';
+import SettingsTabWhite from '../../../../../third_party/icons/settings-tab-white.svg';
 import {
   SiteBoundaries,
   Chips,
